requirements.txt
fastapi[all]
orjson==3.8.0
pydantic==1.10.2
uvicorn[standart]
uvloop==0.17.0; sys_platform == "linux" or sys_platform == "darwin"  
pydantic[dotenv]

SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
asyncpg==0.26.0
psycopg2-binary==2.9.3
databases[postgresql]==0.6.1 

alembic==1.12.0

pytest==7.4.2
pytest-env==1.0.1

python-multipart==0.0.5
python-jose[cryptographic]==3.3.0
passlib[bcrypt]==1.7.4

boto3
python-dotenv
aiofiles
cachetools